+++ packages/shared/widget-plugin-filtering/src/controllers/NumberFilterController.ts
@@ -1,7 +1,7 @@
 import { debounce } from "@mendix/widget-plugin-platform/utils/debounce";
 import { action, autorun, computed, makeObservable, reaction, runInAction } from "mobx";
 import { createRef } from "react";
-import { InputStore } from "../stores/InputStore";
+import { InputStore } from "../stores/input/InputStore";
 import { FilterFunctionBinary, FilterFunctionGeneric, FilterFunctionNonValue } from "../typings/FilterFunctions";
 import { FilterV, Number_InputFilterInterface } from "../typings/InputFilterInterface";
