+++ packages/shared/widget-plugin-filtering/src/controllers/StaticFilterController.ts
@@ -2,7 +2,7 @@ import { executeAction } from "@mendix/widget-plugin-platform/framework/execute-
 import { ActionValue, DynamicValue, EditableValue } from "mendix";
 import { action, autorun, computed, makeObservable, observable, reaction } from "mobx";
 import { OptionsSerializer } from "../stores/OptionsSerializer";
-import { StaticSelectFilterStore } from "../stores/StaticSelectFilterStore";
+import { StaticSelectFilterStore } from "../stores/picker/StaticSelectFilterStore";
 import { OptionWithState } from "../typings/BaseSelectStore";
 
 interface CustomOption<T> {
