+++ packages/shared/widget-plugin-filtering/src/controllers/StringFilterController.ts
@@ -1,7 +1,7 @@
 import { debounce } from "@mendix/widget-plugin-platform/utils/debounce";
 import { action, autorun, computed, makeObservable, reaction, runInAction } from "mobx";
 import { createRef } from "react";
-import { InputStore } from "../stores/InputStore";
+import { InputStore } from "../stores/input/InputStore";
 import {
     FilterFunctionBinary,
     FilterFunctionGeneric,
