+++ packages/shared/widget-plugin-filtering/src/providers/LegacyPv.ts
@@ -2,10 +2,10 @@ import { ListAttributeValue } from "mendix";
 import { FilterCondition } from "mendix/filters";
 import { action, comparer, computed, makeObservable, observable, reaction } from "mobx";
 import { FilterType as Ft, LegacyProvider } from "../context";
-import { DateInputFilterStore } from "../stores/DateInputFilterStore";
-import { NumberInputFilterStore } from "../stores/NumberInputFilterStore";
-import { StaticSelectFilterStore } from "../stores/StaticSelectFilterStore";
-import { StringInputFilterStore } from "../stores/StringInputFilterStore";
+import { DateInputFilterStore } from "../stores/input/DateInputFilterStore";
+import { NumberInputFilterStore } from "../stores/input/NumberInputFilterStore";
+import { StringInputFilterStore } from "../stores/input/StringInputFilterStore";
+import { StaticSelectFilterStore } from "../stores/picker/StaticSelectFilterStore";
 import { InputFilterInterface } from "../typings/InputFilterInterface";
 import { PickerFilterStore } from "../typings/PickerFilterStore";
 import { FiltersSettingsMap } from "../typings/settings";
