--- packages/shared/widget-plugin-filtering/src/stores/IJSActionsControlled.ts
@@ -1,4 +1,0 @@
-export interface IJSActionsControlled {
-    handleResetValue: (useDefaultValue: boolean) => void;
-    handleSetValue: (useDefaultValue: boolean, params: { stringValue: string }) => void;
-}
