+++ packages/shared/widget-plugin-filtering/src/stores/generic/OptionsSerializer.ts
