+++ packages/shared/widget-plugin-filtering/src/stores/input/DateInputFilterStore.ts
@@ -15,13 +15,13 @@ import {
     or
 } from "mendix/filters/builders";
 import { action, comparer, IReactionDisposer, makeObservable, observable, reaction } from "mobx";
-import { betweenToState, isAnd, isEmptyExp, isNotEmptyExp, isOr, singularToState } from "../condition-utils";
-import { FilterFunctionBinary, FilterFunctionGeneric, FilterFunctionNonValue } from "../typings/FilterFunctions";
-import { Date_InputFilterInterface } from "../typings/InputFilterInterface";
-import { FilterFunction } from "../typings/mendix";
-import { FilterData, InputData } from "../typings/settings";
+import { betweenToState, isAnd, isEmptyExp, isNotEmptyExp, isOr, singularToState } from "../../condition-utils";
+import { FilterFunctionBinary, FilterFunctionGeneric, FilterFunctionNonValue } from "../../typings/FilterFunctions";
+import { Date_InputFilterInterface } from "../../typings/InputFilterInterface";
+import { FilterFunction } from "../../typings/mendix";
+import { FilterData, InputData } from "../../typings/settings";
+import { BaseInputFilterStore } from "../BaseInputFilterStore";
 import { DateArgument } from "./Argument";
-import { BaseInputFilterStore } from "./BaseInputFilterStore";
 
 type DateFns = FilterFunctionGeneric | FilterFunctionNonValue | FilterFunctionBinary;
 type StateTuple = [DateFns, Date | undefined, Date | undefined];
