+++ packages/shared/widget-plugin-filtering/src/stores/input/NumberInputFilterStore.ts
@@ -2,13 +2,13 @@ import { Big } from "big.js";
 import { ListAttributeValue } from "mendix";
 import { FilterCondition } from "mendix/filters";
 import { action, comparer, makeObservable } from "mobx";
-import { inputStateFromCond } from "../condition-utils";
-import { FilterFunctionBinary, FilterFunctionGeneric, FilterFunctionNonValue } from "../typings/FilterFunctions";
-import { Number_InputFilterInterface } from "../typings/InputFilterInterface";
-import { FilterData, InputData } from "../typings/settings";
+import { inputStateFromCond } from "../../condition-utils";
+import { FilterFunctionBinary, FilterFunctionGeneric, FilterFunctionNonValue } from "../../typings/FilterFunctions";
+import { Number_InputFilterInterface } from "../../typings/InputFilterInterface";
+import { FilterData, InputData } from "../../typings/settings";
+import { BaseInputFilterStore } from "../BaseInputFilterStore";
+import { baseNames } from "../fn-mappers";
 import { NumberArgument } from "./Argument";
-import { BaseInputFilterStore } from "./BaseInputFilterStore";
-import { baseNames } from "./fn-mappers";
 
 type NumFns = FilterFunctionGeneric | FilterFunctionNonValue | FilterFunctionBinary;
 type Formatter = ListAttributeValue<Big>["formatter"];
