+++ packages/shared/widget-plugin-filtering/src/stores/picker/SearchStore.ts
