+++ packages/shared/widget-plugin-filtering/src/stores/picker/SelectedItemsStore.ts
